+++ data-exporter/src/main/java/org/reactome/release/dataexport/DataExportUtilities.java
@@ -1,4 +1,4 @@
-package org.reactome.release;
+package org.reactome.release.dataexport;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
