+++ data-exporter/src/main/java/org/reactome/release/dataexport/NCBIGene.java
@@ -1,4 +1,4 @@
-package org.reactome.release;
+package org.reactome.release.dataexport;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -11,7 +11,7 @@ import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.*;
 
-import static org.reactome.release.DataExportUtilities.*;
+import static org.reactome.release.dataexport.DataExportUtilities.*;
 
 /**
  * File generator for NCBI Gene.  This class has logic for producing a file for
