+++ data-exporter/src/test/java/org/reactome/release/dataexport/PathwayHierarchyUtilitiesTest.java
@@ -1,4 +1,4 @@
-package org.reactome.release;
+package org.reactome.release.dataexport;
 
 import org.junit.jupiter.api.*;
